Da finire a casa

+++ src/main/java/org/cirmmp/spring/dao/FileListDao.java
@@ -4,15 +4,16 @@ import org.cirmmp.spring.model.FileList;
 
 import java.util.List;
 
+
 public interface FileListDao {
 
-    FileList findById(int id);
+    FileList findById(Long id);
 
   //  List<FileList> findByProjectId(int project_id);
 
     void save(FileList fileList);
 
-    void deleteById(int fillist_id);
+    void deleteById(Long fillist_id);
 
     List<FileList> findAllFiles();
 }
